separate interface from implementation
