+++ posthog/clickhouse/materialized_columns/__init__.py
@@ -1,10 +1,80 @@
-from posthog.models.property import TableWithProperties
+from abc import ABC, abstractmethod
+from datetime import timedelta
+
+from posthog.cache_utils import cache_for
+from posthog.models.property import PropertyName, TableColumn, TableWithProperties
 from posthog.settings import EE_AVAILABLE
+
 
 ColumnName = str
 TablesWithMaterializedColumns = TableWithProperties
 
+
+class MaterializedColumnBackend(ABC):
+    @abstractmethod
+    def get_materialized_columns(
+        self,
+        table: TablesWithMaterializedColumns,
+    ) -> dict[tuple[PropertyName, TableColumn], ColumnName]:
+        raise NotImplementedError
+
+    @abstractmethod
+    def materialize(
+        self,
+        table: TableWithProperties,
+        property: PropertyName,
+        column_name=None,
+        table_column: TableColumn = "properties",
+        create_minmax_index=False,
+    ) -> None:
+        raise NotImplementedError
+
+    @abstractmethod
+    def backfill_materialized_columns(
+        self,
+        table: TableWithProperties,
+        properties: list[tuple[PropertyName, TableColumn]],
+        backfill_period: timedelta,
+        test_settings=None,
+    ) -> None:
+        raise NotImplementedError
+
+
+class DummyMaterializedColumnBackend(MaterializedColumnBackend):
+    def get_materialized_columns(
+        self,
+        table: TablesWithMaterializedColumns,
+    ) -> dict[tuple[PropertyName, TableColumn], ColumnName]:
+        return {}
+
+    def materialize(
+        self,
+        table: TableWithProperties,
+        property: PropertyName,
+        column_name=None,
+        table_column: TableColumn = "properties",
+        create_minmax_index=False,
+    ) -> None:
+        pass
+
+    def backfill_materialized_columns(
+        self,
+        table: TableWithProperties,
+        properties: list[tuple[PropertyName, TableColumn]],
+        backfill_period: timedelta,
+        test_settings=None,
+    ) -> None:
+        pass
+
+
 if EE_AVAILABLE:
-    from ee.clickhouse.materialized_columns.columns import *
+    from ee.clickhouse.materialized_columns.columns import EnterpriseMaterializedColumnBackend
+
+    backend: MaterializedColumnBackend = EnterpriseMaterializedColumnBackend()
 else:
-    from .column import *
+    backend = DummyMaterializedColumnBackend()
+
+
+get_materialized_columns = cache_for(timedelta(minutes=15))(backend.get_materialized_columns)
+materialize = backend.materialize
+backfill_materialized_columns = backend.backfill_materialized_columns
